Blinkup Timeout Causing Issues (iOS).
Cause: Device Poller isn't initialized until the blinkup process is done, and setting the timeout at that point causes issues for subsequent blinkups.
Device Poller is likely intented for custom blinkup views, otherwise it is created by the controller itself. It appears that changing its properties while it is in use is affecting blinkupu stability on ios.

Fix: Removing the logic which sets timeout during blinkup. Note: we will be using the default timeout value for the time being in iOS.
Tested:
- Blinkup / Clear / Blinkup
- Blinkup / Blinkup / Blinkup

Results:
- Not 100% but alot closer, alot fewer timeouts than we were used to.